refactor: move _factory in __init__
